+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.0.3] - 2025-08-17
+### Added
+- Add GetTarget method on DirectionMap.
+
+### Changed
+- Removed int Target property on DirectionMap.
+
+### Fixed
+- Some cs doc comments issues.
+
 ## [0.0.2] - 2025-08-17
 ### Added
 - Samples SharedAssets and APIPlayground.